requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.7
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flack==1.3.7
Flask==2.3.2
Flask-SocketIO==5.3.4
gevent==22.10.2
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-translate==3.11.1
googleapis-common-protos==1.59.0
googletrans==3.0.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
http3==0.6.7
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.6.0
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
msgpack==1.0.5
newsapi-python==0.2.7
oauthlib==3.2.2
onesocial==1.0.1
onesocial-django==1.0.0
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.1
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-engineio==4.4.1
python-socketio==5.8.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-async==0.6.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
routing==0.2.0
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social==0.1.dev3
social-auth-app-django==5.2.0
social-auth-core==4.4.1
soupsieve==2.4.1
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
websocket==0.2.1
Werkzeug==2.3.4
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
